Consolidate duplicated date formatting into src/lib/format.ts (#526)
Replace 5 local formatTimeAgo/formatRelativeDate implementations and 2
duplicated formatFutureDate functions with shared formatRelativeTime and
new formatFutureTime from src/lib/format.ts.
