+++ src/components/settings/pages/BrokenFeedsSettingsContent.tsx
@@ -12,7 +12,7 @@ import { useQueryClient } from "@tanstack/react-query";
 import { toast } from "sonner";
 import { trpc } from "@/lib/trpc/client";
 import { handleSubscriptionDeleted } from "@/lib/cache";
-import { getFeedDisplayName } from "@/lib/format";
+import { getFeedDisplayName, formatRelativeTime, formatFutureTime } from "@/lib/format";
 import { Button, Alert } from "@/components/ui";
 import { SettingsListSkeleton } from "@/components/settings";
 import { UnsubscribeDialog } from "@/components/feeds/UnsubscribeDialog";
@@ -30,70 +30,6 @@ interface BrokenFeed {
   lastError: string | null;
   lastFetchedAt: Date | null;
   nextFetchAt: Date | null;
-}
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-/**
- * Format a date relative to now (e.g., "2 hours ago", "Jan 15, 2024")
- */
-function formatRelativeDate(date: Date | null): string {
-  if (!date) return "Never";
-
-  const now = new Date();
-  const dateObj = new Date(date);
-  const diffMs = now.getTime() - dateObj.getTime();
-  const diffMins = Math.floor(diffMs / (1000 * 60));
-  const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-  const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-
-  if (diffMins < 1) {
-    return "Just now";
-  } else if (diffMins < 60) {
-    return `${diffMins} minute${diffMins === 1 ? "" : "s"} ago`;
-  } else if (diffHours < 24) {
-    return `${diffHours} hour${diffHours === 1 ? "" : "s"} ago`;
-  } else if (diffDays === 1) {
-    return "Yesterday";
-  } else if (diffDays < 7) {
-    return `${diffDays} days ago`;
-  } else {
-    return dateObj.toLocaleDateString("en-US", {
-      month: "short",
-      day: "numeric",
-      year: "numeric",
-    });
-  }
-}
-
-/**
- * Format a future date (for next fetch time)
- */
-function formatFutureDate(date: Date | null): string {
-  if (!date) return "Not scheduled";
-
-  const now = new Date();
-  const dateObj = new Date(date);
-  const diffMs = dateObj.getTime() - now.getTime();
-
-  // If in the past, it's scheduled to run soon
-  if (diffMs <= 0) {
-    return "Pending";
-  }
-
-  const diffMins = Math.floor(diffMs / (1000 * 60));
-  const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-  const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-
-  if (diffMins < 60) {
-    return `In ${diffMins} minute${diffMins === 1 ? "" : "s"}`;
-  } else if (diffHours < 24) {
-    return `In ${diffHours} hour${diffHours === 1 ? "" : "s"}`;
-  } else {
-    return `In ${diffDays} day${diffDays === 1 ? "" : "s"}`;
-  }
 }
 
 // ============================================================================
@@ -287,8 +223,11 @@ function BrokenFeedRow({ feed, onUnsubscribe }: BrokenFeedRowProps) {
               {feed.consecutiveFailures} consecutive failure
               {feed.consecutiveFailures === 1 ? "" : "s"}
             </span>
-            <span>Last attempt: {formatRelativeDate(feed.lastFetchedAt)}</span>
-            <span>Next retry: {formatFutureDate(feed.nextFetchAt)}</span>
+            <span>
+              Last attempt:{" "}
+              {feed.lastFetchedAt ? formatRelativeTime(new Date(feed.lastFetchedAt)) : "Never"}
+            </span>
+            <span>Next retry: {formatFutureTime(feed.nextFetchAt)}</span>
           </div>
         </div>
